clase-08/testing-app/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
mysql-connector-python==8.3.0
PyMySQL==1.1.0
Flask-Cors==4.0.0
Requests==2.31.0